Add some more testcases
