+++ packages/integration-tests/fixtures/array-entries-option/input/entrypoint1.js
