+++ packages/integration-tests/fixtures/array-entries-option/input/entrypoint2.js
