+++ packages/integration-tests/fixtures/array-entries-option/input/entrypoint3.js
