+++ packages/integration-tests/fixtures/basic-release-injection/setup.ts
@@ -4,4 +4,7 @@ import { createCjsBundles } from "../../utils/create-cjs-bundles";
 const entryPointPath = path.resolve(__dirname, "./input/entrypoint.js");
 const outputDir = path.resolve(__dirname, "./out");
 
-createCjsBundles({ index: entryPointPath }, outputDir, { release: "I AM A RELEASE!", include: "" });
+createCjsBundles({ index: entryPointPath }, outputDir, {
+  release: "I AM A RELEASE!",
+  include: outputDir,
+});
