+++ packages/integration-tests/fixtures/regex-entries-option/input/entrypoint1.js
@@ -1,0 +1,2 @@
+// eslint-disable-next-line @typescript-eslint/no-unsafe-argument, @typescript-eslint/no-unsafe-member-access
+process.stdout.write((global.SENTRY_RELEASE || {}).id || "");
